Move planner data into zustand store
