+++ apps/antalmanac/src/components/RightPane/CoursePane/CoursePaneRoot.tsx
@@ -1,23 +1,20 @@
-import { Box } from '@mui/material';
-import { usePostHog } from 'posthog-js/react';
-import { useCallback, useEffect } from 'react';
-
 import { CoursePaneButtonRow } from '$components/RightPane/CoursePane/CoursePaneButtonRow';
 import CourseRenderPane from '$components/RightPane/CoursePane/CourseRenderPane';
 import { SearchForm } from '$components/RightPane/CoursePane/SearchForm/SearchForm';
 import RightPaneStore from '$components/RightPane/RightPaneStore';
-import { usePlannerRoadmaps } from '$hooks/usePlanner';
 import analyticsEnum, { logAnalytics } from '$lib/analytics/analytics';
 import { Grades } from '$lib/grades';
 import { WebSOC } from '$lib/websoc';
 import { useCoursePaneStore } from '$stores/CoursePaneStore';
 import { openSnackbar } from '$stores/SnackbarStore';
+import { Box } from '@mui/material';
+import { usePostHog } from 'posthog-js/react';
+import { useCallback, useEffect } from 'react';
 
 export function CoursePaneRoot() {
     const { key, forceUpdate, searchFormIsDisplayed, displaySearch, displaySections, advancedSearchEnabled } =
         useCoursePaneStore();
     const postHog = usePostHog();
-    usePlannerRoadmaps();
 
     const handleSearch = useCallback(() => {
         if (!advancedSearchEnabled) {
