+++ apps/antalmanac/src/lib/helpers.ts
@@ -1,6 +1,5 @@
-import { MouseEvent } from 'react';
-
 import { openSnackbar } from '$stores/SnackbarStore';
+import { MouseEvent } from 'react';
 
 export const warnMultipleTerms = (terms: Set<string>) => {
     openSnackbar(
@@ -17,3 +16,10 @@ export async function clickToCopy(event: MouseEvent<HTMLElement>, sectionCode: s
 }
 
 export const FAKE_LOCATIONS = ['VRTL REMOTE', 'ON LINE', 'TBA'];
+
+export const QUARTER_ORDER_IN_YEAR: Record<string, number> = {
+    Winter: 0,
+    Spring: 1,
+    Summer: 2,
+    Fall: 3,
+};
